pyproject.toml
[tool.poetry]
name = "graia-spinach-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "graia_spinach_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
graia-ariadne = "^0.11.0"
graia-saya = "^0.0.17"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
